update CopyElements (merge request !394)

+++ 3rdparty/LLM_kernels/csrc/kernels/nvidia/samplers/copy_elements.cuh
@@ -10,7 +10,7 @@ namespace llm_kernels {
 namespace nvidia {
 
 template <typename T>
-void InvokeCopyElements(T** src_ptrs, T* dest, u_int64_t num_elements, cudaStream_t& stream);
+void InvokeCopyElements(T** src_ptrs, T* dest, size_t num_elements, cudaStream_t& stream);
 
 }  // namespace nvidia
 }  // namespace llm_kernels
